hnswrag/requirements.txt
numpy>=1.24,<2.0
tqdm>=4.66
pypdf>=4.0.0
sentence-transformers>=2.6.1
torch>=2.1.0
faiss-cpu>=1.7.4
rank-bm25>=0.2.2